backend/requirements.txt
Django==4.2.4
django-filter==23.2
djangorestframework==3.14.0
djoser==2.2.0
drf-extra-fields==3.7.0
Pillow==10.0.0
psycopg2-binary==2.9.7
python-dotenv==0.21.0
gunicorn==20.1.0